4897: Close issue https://github.com/mozilla-mobile/android-components/issues/4876: Add UI for addon discovery and installation. r=csadilek a=Amejia481

[video](https://drive.google.com/file/d/13YqFm7uYh5Tk_h-l99Ca0HcCxQSBxloE/view?usp=sharing)

+++ mobile/android/android-components/components/feature/addons/src/main/java/mozilla/components/feature/addons/amo/AddOnsCollectionsProvider.kt
@@ -87,7 +87,7 @@ internal fun JSONObject.getRating(): AddOn.Rating? {
     return if (jsonRating != null) {
         AddOn.Rating(
             reviews = jsonRating.optInt("count"),
-            average = jsonRating.optDouble("average")
+            average = jsonRating.optDouble("average").toFloat()
         )
     } else {
         null
